day23

+++ internal/model/expense.go
@@ -1,9 +1,9 @@
 package model
 
 type Expense struct {
-	ID       int    `json:"id"`
-	Title    string `json:"title"`
-	Amount   int    `json:"amount"`
-	Category string `json:"category"`
-	UserID   int    `json:"user_id"`
+	ID       int     `json:"id"`
+	Title    string  `json:"title"`
+	Amount   float64 `json:"amount"`
+	Category string  `json:"category"`
+	UserID   int     `json:"user_id"`
 }
